Update searchform.php
role="search" can not be placed on a ```<form>``` element, as it will overwrite the form's inherent role.  Instead, the search role should be applied to a ```<div>``` that wraps all the elements that make up the search component, within the ```<form>```.

Inputs should always have an a label associated with them. To rectify this, I've added an ```aria-label="Search:" to stand in for the label element.  

Removed the ```role="button"``` from the submit button, as it's redundant to the inherent semantics of the button element.